Updated programs
Hi Neena,

The only thing I want to point out is that when recoding extreme values for visits, there's 2 changed in US and 1 changed in Italy which was different from the comment you sent me via email. I used data received on 3/2 so this might be different again with the newest data.

Thanks,
Mia